src/preprocess_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-multipart>=0.0.6
Pillow>=9.5.0
numpy>=1.24.0
prometheus-client>=0.16.0
python-multipart>=0.0.6